Install Adobe Reader Script
I was tasked with building and deploying 130 workstations running Windows 10 Enterprise, is it possible?? Yes! It was, because I was able to automate several tasks that would consume time and delay production. Ever install Adobe Reader? Its not as easy as installing a browser, you have to be aware of which version you're installing and additional add-on features. When you download Adobe Reader it will automatically check "install Google Chrome/McAfee AV" which could wreak havoc and almost did when one of my technicians thought it'd be cool to install those while Adobe was installing. Since then, I created this script that will set execution policies to allow executable permissions with remote signatures. Upon completion, Adobe reader will be fully installed! :)